ITensors support

This package also supports convertion with the Site and OpSum objects from ITensors library and management of DMRG sweeping process. To use these functions, include

using ITensors

at the heading of the Julia script.

Format conversion

The Sites objects in ITensors can be converted to a Confs object with the conserved quantities extracted.

Main.FuzzifiED.ConfsFromSites — Method

function ConfsFromSites(sites :: Vector{Index{Vector{Pair{QN, Int64}}}}, qn_s :: QN) :: Confs

Converts a Sites object in the ITensors package to the Confs object

Arguments

sites :: Vector{Index{Vector{Pair{QN, Int64}}}} is a Sites object. Only Fermion site type is supported, and the quantum numbers of the 0 state must be all zero. Note that this will subject to the limitation in ITensors that the number of conserved quantities must be less than 4.

qn_s :: QN is a QN object that specifies the the quantum number of the selected configuration.

Main.FuzzifiED.ConfsFromSites — Method

function ConfsFromSites(sites :: Vector{Index{Vector{Pair{QN, Int64}}}}, cf_ref :: Vector{Int64}) :: Confs

sites :: Vector{Index{Vector{Pair{QN, Int64}}}} is a Sites object, where the modulus-$\pm 1$ quantum numbers will be taken out, and QNs with other modulus will be discarded automatically. Also note that only Fermion site type is supported, and the quantum numbers of the 0 state must be all zero. Note that this will subject to the limitation in ITensors that the number of conserved quantities must be less than 4.

cf_ref :: Vector{Int64}) is a reference configuration composed of 0 and 1 in ITensors format.

Conversely, the Sites objects in ITensors can also be generated from a set of diagonal quantum numbers

Main.FuzzifiED.SitesFromQnu — Method

function SitesFromQnu(; qnu_o :: Vector{Vector{Int64}}, qnu_name :: Vector{String}, modul :: Vector{Int64})

returns the ITensors Sites object from the information of quantum numbers

Arguments

qnu_o :: Vector{Vector{Int64}} stores the charge of each orbital under each conserved quantity. See Confs for detail.

qnu_name :: Vector{String} stores the name of each quantum number. Facultive, QN1, QN2, ... by default.

modul :: Vector{Int64} stores the modulus of each quantum number. Store 1 if no modulus. Facultive, all 1 by default.

If the number of quantum numbers are too many, it can be truncated by

Main.FuzzifiED.TruncateQnu — Method

TruncateQnu(; qnu_o :: Vector{Vector{Int64}}, qnu_name :: Vector{String}, modul :: Vector{Int64}, trunc_lth :: Int64, trunc_wt :: Vector{Int64})

truncates the list of $N_U$ QNU's from to a number $N'_U$ acceptable by ITensors. The new quantum numbers are

\[ Q'_1=Q_1,\ Q'_2=Q_2,\ \dots,\ Q'_{N'_U-1}=Q_{N'_U-1}\ Q'_{N'_U}=\lambda_{N'_U}Q_{N'_U}+\lambda_{N'_U+1}Q_{N'_U+1}+\dots+\lambda_{N_U}Q_{N_U}\]

Arguments

qnu_o :: Vector{Vector{Int64}} stores the charge of each orbital under each conserved quantity. See Confs for detail.

qnu_name :: Vector{String} stores the name of each quantum number. Facultive, QN1, QN2, ... by default.

modul :: Vector{Int64} stores the modulus of each quantum number. Store 1 if no modulus. Facultive, all 1 by default.

trunc_lth :: Int64 stores the truncated numbers of QNU. Facultive, 3 by default.

trunc_wt :: Vecotr{Int64} stores the $N_U-N'_U+1$ coefficients $\lambda$. Facultive, $1,2,4,8,\dots$ by default.

Output

A named tuple with three elements that can be directly fed into SitesFromQnu

qnu_o :: Vector{Vector{Int64}} stores the charge of each orbital under each conserved quantity. See Confs for detail.

qnu_name :: Vector{String} stores the name of each quantum number.

modul :: Vector{Int64} stores the modulus of each quantum number, 1 if no modulus.

The OpSum objects in ITensors can be converted with the collection of Term's

Main.FuzzifiED.TermsFromOpSum — Method

function TermsFromOpSum(opsum :: Sum{Scaled{ComplexF64, Prod{Op}}}) :: Vector{Term}

Converts a OpSum object in ITensors to a series of terms. Note that the only operators supported are "C", "Cdag" "N" and "I".

Main.FuzzifiED.OpSumFromTerms — Method

function OpSumFromTerms(tms :: Vector{Term}) :: Sum{Scaled{ComplexF64, Prod{Op}}}

Converts a series of terms to OpSum object in ITensors.

Easy sweep

This tool kit facilitates the management of DMRG process. It automatically records the intermediate results and recover these results if a job is stopped and run again on HPC. It also manages the gradual increase of maximal bond dimensions and the determination of convergence by the criteria of energy. These functions require the package ITensorMPOConstruction, which can be installed by

julia> using Pkg; Pkg.add(url="https://github.com/ITensor/ITensorMPOConstruction.jl.git")

Main.FuzzifiED.EasySweep — Method

function EasySweep(id :: String, hmt :: MPO, st00 :: MPS ; path :: String, dim_list :: Vector{Int64}, proj :: Vector{String}, e_tol1 :: Float64, e_tol :: Float64, cutoff :: Vector{Float64}, maxdim0 :: Vector{Float64}, noise0 :: Vector{Float64}, noise :: Vector{Int64}, nsweeps :: Int64, weight :: Float64, observer :: AbstractObserver) :: Tuple{Float64, MPS}

Function

This function automatically performs several rounds of DMRG sweeps with increasing bond dimensions. It first checks the file st_$(id).h5 in a specified repository. If the key st_fin exists, it reads the energy and MPS from the file and return the energy and MPS, otherwise it will perform the DMRG process. For each round, it will try to access the results from the key st_d$(dim_i) in st_$(id).h5, where dim_i is either 0 representing the initial round, or an element of array dim_list. If the key exist, it will read the result ; otherwise it will perform the sweeps using SweepOne. For the initial round, it will take the initial state from st00, the maximal bond dimensions from maxdim0, noise from noise0 and record the results in the key E_d0 and st_d0 in st_$(id).h5. For each of the following round, it will take the result from the previous round as the initial state and perform nsweeps sweeps with the bond dimension dim_list[i]. Each round will be stopped if the energy difference is less than e_tol1. The entire process will be stopped if the energy difference between two rounds is less than e_tol or the bond dimension of the result is less than 0.9 times the maximal bond dimension. The projected states will be accessed from the files specified by proj. It will try to access first the states with the same bond dimension as the projected states. If such states do not exist, it will then access the final state. The resulting energy will be written into the key E_fin in the file st_$(id).h5, and the MPS written into st_fin. The function returns a tuple of energy and the final MPS.

Arguments

id :: String is a string identifying the file to which the results will be accessed and written.

hmt :: MPO is an MPO specifying the Hamiltonian.

st00 :: MPS is an MPS specifying the initial state.

path :: String identifies the path where the results will be accessed and stored. Facultive, ./ by default.

dim_list :: Vector{Int64} :: Int64 is a list that specifies the maximal bond dimensions of each round of sweeps starting from the second round. Facultive, [1000,2000,3000,4000,5000,6000] by default

proj :: Vector{String} specifies the name of the states that will be projected. Facultive, empty by default.

e_tol1 :: Float64 specifies the energy tolerence as a criteria to end the round of sweeps for each round of sweeps. Facultive, 1E-6 by default.

e_tol :: Float64 specifies the energy tolerence as a criteria to end the entire process. Facultive, 1E-7 by default.

cutoff :: Vector{Float64} is the cutoff that will be sent into DMRG. Facultive, [1.E-9] by default.

maxdim0 :: Vector{Int64} specifies the maximal bond dimensions of the first round of sweeps. Facultive, [1E-4,3E-5,1E-5,3E-6,1E-6,3E-7] by default.

noise0 :: Vector{Float64} specifies the noise of each sweep in the initial round and will be sent into DMRG. Facultive, [1E-4,3E-5,1E-5,3E-6,1E-6,3E-7] by default.

noise :: Vector{Float64} specifies the noise of each sweep from the second round and will be sent into DMRG. Facultive, [1E-6,1E-7,0] by default.

nsweeps :: Int64 specifies the number of sweeps in each round from the second rounds. Facultive, 10 by default.

weight :: Float64 specifies the weight of projected states and will be sent into DMRG. Facultive, 100.0 by default.

observer :: AbstractObserver specifies the measurement and cutoff condition for each sweep starting from the second round. Facultive, by default the observer will print the energy and cutoff once the energy difference is less than e_tol at each sweep.

Main.FuzzifiED.SweepOne — Method

function SweepOne(id :: String, hmt :: MPO, st0 :: MPS, dim1 :: Int64 ; path :: String, cutoff :: Vector{Float64}, maxdim :: Vector{Int64}, nsweeps :: Int64, noise :: Vector{Float64}, proj :: Vector{String}, e_tol :: Float64, weight :: Float64, observer :: AbstractObserver) :: Tuple{Float64, MPS}

Function

This function performs one round of nsweeps sweeps. It first checks the file st_$(id).h5 in a specified repository. If the key st_d$(dim1) exists, it reads the energy and MPS from the file and return the energy and MPS, otherwise it will perform the DMRG process with the maximal bond dimension specified by maxdim if it exists, or dim1. The projected states will be read from the key st_d$(dim1) if it exists or st_fin in the file st_$(fi).h5 in the same repository for each string fi in the array proj. The sweeps will be ended if the energy difference is less than etol or whatever criteria is given in observer. The resulting energy will be written into the key E_d$(dim1) in the file st_$(id).h5, and the MPS written into st_d$(dim1). The function returns a tuple of energy and the final MPS.

Arguments

id :: String is a string identifying the file to which the results will be accessed and written.

hmt :: MPO is an MPO specifying the Hamiltonian.

st0 :: MPS is an MPS specifying the initial state.

dim1 :: Int64 is a bond dimension that will be used to identify the result.

path :: String identifies the path where the results will be accessed and stored. Facultive, ./ by default.

cutoff :: Vector{Float64} is the cutoff that will be sent into DMRG. Facultive, [1.E-9] by default.

maxdim :: Vector{Int64} specifies the maximal bond dimension of each sweep. Facultive, [dim1] by default.

nsweeps :: Int64 specifies the number of sweeps in the round. Facultive, 10 by default.

noise :: Vector{Float64} specifies the noise of each sweep and will be sent into DMRG. Facultive, [1E-6,1E-7,0] by default.

proj :: Vector{String} specifies the name of the states that will be projected. Facultive, empty by default.

e_tol :: Float64 specifies the energy tolerence as a criteria to end the sweeps. Facultive, 1E-6 by default.

weight :: Float64 specifies the weight of projected states and will be sent into DMRG. Facultive, 100.0 by default.

observer :: AbstractObserver specifies the measurement and cutoff condition for each sweep. Facultive, by default the observer will print the energy and cutoff once the energy difference is less than e_tol at each sweep.

Main.FuzzifiED.GetMpoSites — Method

function GetMpoSites(id :: String, tms :: Union{Vector{Term}, Sum{Scaled{ComplexF64, Prod{Op}}}} ; path :: String, qnu_o :: Vector{Vector{Int64}}, qnu_name :: Vector{String}, modul :: Vector{Int64}, mpo_method :: Function) :: Tuple{MPO, Vector{<:Index}}

Function

This function returns the MPO and sites for a given operator and a Hilbert space with given quantum numbers. It first checks the file op_$(id).h5 in a specified repository. If the file exists, it will try to read the fields mpo and sites and return the MPO and Sites. Otherwise it will first generates the sites with the quantum numbers given in qnu_o, qnu_name and modul (these objects are often results of a function named Get*Qnu). Then it will generate the MPO with the terms of the operator given in tms. The MPO and sites will be written into the file op_$(id).h5 in the fields mpo and sites.

Arguments

id :: String is a string identifying the file to which the results will be accessed and written.

tms :: Vector{Term} or tms :: Sum{Scaled{ComplexF64, Prod{Op}}} is either an array of terms or a OpSum objects that specifies the expression of the operator.

path :: String identifies the path where the results will be accessed and stored. Facultive, ./ by default.

qnu_o :: Vector{Vector{Int64}} is a list where each element specifies a quantum number. Each element is a list that specifies the charges of each orbital under the quantum number. Obligatory.

qnu_name :: Vector{String} specifies the name of each quantum number. Facultive, QN1, QN2, ... by default.

module :: Vector{Int64} specifies the modulus of each quantum number. Facultive, all 1 by default.

mpo_method :: Function is a function mpo_method(os :: OpSum, sites :: Sites) :: MPO that generates the MPO from OpSum and Sites. Facultive, MPO by default. We suggest using MPO_new in ITensorMPOConstruction package. See example_ising_dmrg_easysweep.jl for example.

Main.FuzzifiED.GetMpo — Method

function GetMpo(id :: String, tms :: Union{Vector{Term}, Sum{Scaled{ComplexF64, Prod{Op}}}} ; path :: String, qnu_o :: Vector{Vector{Int64}}, qnu_name :: Vector{String}, modul :: Vector{Int64}, mpo_method :: Function) :: Tuple{MPO, Vector{<:Index}}

Function

This function returns the MPO for a given operator and a given set of sites. It first checks the file op_$(id).h5 in a specified repository. If the file exists, it will try to read the fields mpo and return the MPO it has read. Otherwise it will generate the MPO with the terms of the operator given in tms. The MPO and Sites will be written into the file op_$(id).h5 in the fields mpo.

Arguments

id :: String is a string identifying the file to which the results will be accessed and written.

tms :: Vector{Term} or tms :: Sum{Scaled{ComplexF64, Prod{Op}}} is either an array of terms or a OpSum objects that specifies the expression of the operator.

sites :: Vector{<:Index} specifies the sites that the operator is acting on.

path :: String identifies the path where the results will be accessed and stored. Facultive, ./ by default.

mpo_method :: Function is a function mpo_method(os :: OpSum, sites :: Sites) :: MPO that generates the MPO from OpSum and Sites. Facultive, MPO by default. We suggest using MPO_new in ITensorMPOConstruction package. See example_ising_dmrg_easysweep.jl for example.