Reference

Configurations

Main.FuzzifiED.Confs — Type

Confs

stores all the configurations that respects the U(1) conserved quantities.

Fields

no :: Int64 is the number of orbitals

ncf :: Int64 is the number of configurations

conf :: Vector{Int64} is an array of length ncf containing all the configurations. Each configuration is expressed in a binary number. If the o-1-th bit of conf[i] is 1, then the o-th orbital in the i-th configuration is occupied ; if the bit is 0, then the orbital is empty.

nor :: Int64, lid :: Vector{Int64} and rid :: Vector{Int64} contain the information of Lin table that is used to inversely look up the index i from the configuration.

Main.FuzzifiED.Confs — Method

Confs(no :: Int64, qnu_s :: Vector{Int64}, qnu_o :: Vector{Vector{Int64}} ; nor :: Int64) :: Confs

Arguments

no :: Int64 is the number of orbitals $N_o$ ;

qnu_s :: Vector{Int64} is the set of $Q_i$ for the selected configurations ;

qnu_o :: Vector{Vector{Int64}} is the set of $q_{io}$ for each quantum number and for each orbital. It should contain $N_U$ elements and each element should be a vector of length $N_o$.

(nor :: Int64 is the number of less significant bits used to generate the Lin table.)

Main.FuzzifiED.ITensorSupport.ConfsFromSites — Method

ConfsFromSites(sites :: Vector{Index{Vector{Pair{QN, Int64}}}}, qn_s :: QN)

Arguments

sites :: Vector{Index{Vector{Pair{QN, Int64}}}} is a Sites object, where the modulus-$\pm 1$ quantum numbers will be taken out, and QNs with other modulus will be discarded automatically. Also note that only Fermion site type is supported, and the quantum numbers of the 0 state must be all zero. Note that this will subject to the limitation in ITensors that the number of conserved quantities must be less than 4.

qn_s :: QN is a QN object that specifies the the quantum number of the selected configuration.

Main.FuzzifiED.ITensorSupport.ConfsFromSites — Method

ConfsFromSites(sites :: Vector{Index{Vector{Pair{QN, Int64}}}}, cf_ref :: Vector{Int64})

sites :: Vector{Index{Vector{Pair{QN, Int64}}}} is a Sites object, where the modulus-$\pm 1$ quantum numbers will be taken out, and QNs with other modulus will be discarded automatically. Also note that only Fermion site type is supported, and the quantum numbers of the 0 state must be all zero. Note that this will subject to the limitation in ITensors that the number of conserved quantities must be less than 4.

cf_ref :: Vector{Int64}) is a reference configuration composed of 0 and 1 in ITensors format.

Basis

Main.FuzzifiED.Basis — Type

Basis

Fields

cfs :: Confs is the basis with only conserved quantities generated in the last step ;

dim :: Int64 is the dimension of the basis ;

szz :: Int64 records the maximum size $\max m_g$ of groups;

cfgr :: Vector{Int64} is a vector of length cfs.ncf and records which group $|I\rangle$ each configuration $|i\rangle$ belong to ;

cffac :: Vector{ComplexF64} is a vector of length cfs.ncf and records the coefficients $\lambda_i$ ;

grel :: Matrix{Int64} is a szz$\times$dim matrix that records the configurations in each group $|i_{I1}\rangle,\dots,|i_{Im_I}\rangle$

grsz :: Vector{Int64} is a vector of length dim that records the size $m_I$ of each group.

Main.FuzzifiED.Basis — Method

Basis(cfs :: Confs, qnz_s :: Vector{ComplexF64}, cyc :: Vector{Int64}, perm_o :: Vector{Vector{Int64}}, ph_o :: Vector{Vector{Int64}}, fac_o :: Vector{Vector{ComplexF64}}) :: Basis

Arguments

cfs :: Confs is the configuration set with only conserved quantities generated in the last step ;

qnz_s :: Vector{ComplexF64} is a vector of length the same as the number of discrete symmetries $N_Z$ that records the eigenvalue of each transformation ;

cyc :: Vector{Int64} records the cycle of each transformation. For $\mathbb{Z}_n$ symmetry, record $n$ ;

perm_o :: Vector{Vector{Int64}} records the permutation $\pi_o$. It has $N_Z$ elements and each of its elements is a vector of length $N_o$.

ph_o :: Vector{Vector{Int64}} records $p_o$ to determine whether or not to perform a particle-hole transformation. It has $N_Z$ elements and each of its elements is a vector of length $N_o$.

fac_o :: Vector{Vector{ComplexF64}} records the factor $p_o$ is determine whether or not to perform a particle-hole transformation. It has $N_Z$ elements and each of its elements is a vector of length $N_o$.

Main.FuzzifiED.Basis — Method

Basis(cfs :: Confs) :: Basis

Generate a basis from the configurations without applying the $\mathbb{Z}_2$ symmetries

Operator

Main.FuzzifiED.Term — Type

Term

A Term object records a term that looks like $Uc^{(p_1)}_{o_1}c^{(p_2)}_{o_2}\dots c^{(p_l)}_{o_l}$ in an operator

Fields

coeff :: ComplexF64 records the coefficient $U$

cstr :: Vector{Int64} is a length-$2l$ vector $(p_1,o_1,p_2,o_2,\dots p_l,o_l)$ recording the operator string

Main.FuzzifiED.Operator — Type

Operator

An Operator object records the sum of terms together with information about its symmetry.

Fields

bsd :: Basis is the basis of the initial state ;

bsf :: Basis is the basis of the final state ;

red_q :: Int64 is a flag that records whether or not the conversion to a sparse martrix can be simplified : if bsd and bsf have exactly the same quantum numbers, and the operator fully respects the symmetries, and all the elements in bsd.cffac and bsf.cffac has the same absolute value, then red_q = 1 ; otherwise red_q = 0 ;

sym_q :: Int64 records the symmetry of the operator : if the matrix is Hermitian, then sym_q = 1 ; if it is symmetric, then sym_q = 2 ; otherwise sym_q = 0 ;

ntm :: Int64 is the number of terms ;

nc :: Int64 is the maximum number of operators in an operator string

cstrs :: Matrix{Int64} is a matrix recording the operator string of each term. Each column corresponds to a term and is padded to the maximum length with -1's.

coeffs :: Vector{ComplexF64} corresponds to the coefficients in each term.

Main.FuzzifiED.Operator — Method

Operator(bsd :: Basis, bsf :: Basis, cstr_vec :: Vector{Vector{Int64}}, coeffs :: Vector{ComplexF64} ; red_q :: Int64, sym_q :: Int64) :: Operator

Arguments

bsd :: Basis is the basis of the initial state ;

bsf :: Basis is the basis of the final state ;

terms :: Vector{Term} records the terms ;

red_q :: Int64 is a flag that records whether or not the conversion to a sparse martrix can be simplified : if bsd and bsf have exactly the same quantum numbers, and the operator fully respects the symmetries, and all the elements in bsd.cffac and bsf.cffac has the same absolute value, then red_q = 1 ; otherwise red_q = 0 ;

sym_q :: Int64 records the symmetry of the operator : if the matrix is Hermitian, then sym_q = 1 ; if it is symmetric, then sym_q = 2 ; otherwise sym_q = 0.

Main.FuzzifiED.ITensorSupport.OperatorFromOpSum — Method

OperatorFromOpSum(bsd :: Basis, bsf :: Basis, opsum :: Sum{Scaled{ComplexF64, Prod{Op}}} ; red_q :: Int64 = 0, sym_q :: Int64 = 0)

Note that the only operators supported are "C", "Cdag" "N" and "I".

Base.:* — Method

*(op :: Operator, st_d :: Vector{ComplexF64})

Measure the action of an operator on a state. st_d must be of length op.bsd.dim. Returns a vector of length op.bsf.dim that represents the final state.

Note that sometimes it is needed to transform a state from one basis to another. This can be done by constructing an identity operator.

stf = Operator(bsd, bsf, [[-1, -1]], [ComplexF64(1)]) * std

Base.:* — Method

*(st_fp :: LinearAlgebra.Adjoint{ComplexF64, Vector{ComplexF64}}, op :: Operator, st_d :: Vector{ComplexF64}) :: ComplexF64

Measuring the inner product between two states and an operator. st_d must be of length op.bsd.dim and st_fp must be of length op.bsf.dim, and st_fp must be an adjoint.

Sparse matrix

Main.FuzzifiED.OpMat — Type

OpMat

Fields

dimd :: Int64 and dimf :: Int64 are the number of columns and rows of the matrix ;

symq :: Int64 records whether or not the matrix is Hermitian or symmetric ;

nel :: Int64 records the number of elements ;

colptr :: Vector{Int64}, rowid :: Vector{Int64} and elval :: Vector{ComplexF64} records the elements of the sparse matrix as in the SparseMatrixCSC elements of Julia.

Main.FuzzifiED.OpMat — Method

OpMat(op :: Operator) :: OpMat

Base.:* — Method

*(mat :: OpMat, st_d :: Vector{ComplexF64})

Measure the action of a sparse matrix on a state. st_d must be of length mat.dimd. Returns a vector of length mat.dimf that represents the final state.

Base.:* — Method

*(st_fp :: LinearAlgebra.Adjoint{ComplexF64, Vector{ComplexF64}}, mat :: OpMat, st_d :: Vector{ComplexF64}) :: ComplexF64

Measuring the inner product between two states and a sparse matrix. st_d must be of length mat.dimd and st_fp must be of length mat.dimf, and st_fp must be an adjoint.

Main.FuzzifiED.GetEigensystem — Method

GetEigensystem(mat :: OpMat, nst :: Int64 ; tol :: Float64 = 1E-8)

Output

A length-nst array recording the eigenvalues, and

A dimd$\times$nst matrix where every column records an eigenstate.