Reference

Basics

Configurations

Main.FuzzifiED.Confs — Type

mutable struct Confs

This type stores all the configurations that respects the conserved quantities and also a table to inversely look up the index from the configuration.

Fields

no :: Int64 is the number of orbitals

ncf :: Int64 is the number of configurations

conf :: Vector{Int64} is an array of length ncf containing all the configurations. Each configuration is expressed in a binary number. If the o-1-th bit of conf[i] is 1, then the o-th orbital in the i-th configuration is occupied ; if the bit is 0, then the orbital is empty.

nor :: Int64, lid :: Vector{Int64} and rid :: Vector{Int64} contain the information of Lin table that is used to inversely look up the index i from the configuration.

It can be generated by the following method.

Main.FuzzifiED.Confs — Method

function Confs(no :: Int64, qnu_s :: Vector{Int64}, qnu_o :: Vector{Vector{Int64}} ; nor :: Int64 = div(no, 2), modul :: Vector{Int64}) :: Confs

generates the configurations that has the diagonal quantum numbers given by qnu_s of certain conserved quantities specified by qnu_o :: Vector{Vector{Int64}}

\[Q_i=\sum_{o=1}^{N_o}q_{io}n_o\]

or

\[Q_i=\sum_{o=1}^{N_o}q_{io}n_o\ \mathrm{mod}\ p_i\]

where $i=1,\dots,N_U$ is the index of conserved quantities, $o$ is the index of orbital, $n_o=c^\dagger_oc_o$, and $q_o$ is a set of coefficients that must be non negative integer valued.

Arguments

no :: Int64 is the number of orbitals $N_o$ ;

qnu_s :: Vector{Int64} is the set of $Q_i$ for the selected configurations ;

qnu_o :: Vector{Vector{Int64}} is the set of $q_{io}$ for each quantum number and for each orbital. It should contain $N_U$ elements and each element should be a vector of length $N_o$.

nor :: Int64 is the number of less significant bits used to generate the Lin table. Facultive, $N_o/2$ by default.

modul :: Vector{Int64} is the modulus of each quantum number. Setting it to 1 means there is no modulus. Facultive, all 1 by default.

Output

cfs :: Confs is a Confs object

The inverse look-back for a configuration can be done by

Main.FuzzifiED.GetConfId — Method

GetConfId(cfs :: Confs, cf :: Int64) :: Int64

inversely look up the index from the configuration

Arguments

cfs :: Confs stores the configurations.

cf :: Int64 stores the configuration to be looked-up expressed in a binary number. If the o-1-th bit of conf[i] is 1, then the o-th orbital in the i-th configuration is occupied ; if the bit is 0, then the orbital is empty.

Output

id :: Int64 is the id of the configuration such that cfs.conf[id] == cf

Basis

Main.FuzzifiED.Basis — Type

mutable struct Basis

This type stores the information of the basis that respects both the conserved quantities and the discrete symmetries. The states in the basis is in the form

\[|I\rangle=\lambda_{i_{I1}}|i_{I1}\rangle+\lambda_{i_{I2}}|i_{I2}\rangle+\cdots+\lambda_{i_{Im_I}}|i_{Im_I}\rangle\]

where $|i\rangle$ is a direct product state, i.e., the configurations $|i_{Ik}\rangle$ are grouped into a state $|I\rangle$.

Fields

cfs :: Confs stores the configurations that respect the conserved quantities ;

dim :: Int64 is the dimension of the basis ;

szz :: Int64 records the maximum size $\max m_g$ of groups;

cfgr :: Vector{Int64} is a vector of length cfs.ncf and records which group $|I\rangle$ each configuration $|i\rangle$ belong to ;

cffac :: Vector{ComplexF64} is a vector of length cfs.ncf and records the coefficients $\lambda_i$ of each configuration ;

grel :: Matrix{Int64} is a szz\*dim matrix that records the configurations in each group $|i_{Ik}\rangle (k = 1,\dots,m_I)$

grsz :: Vector{Int64} is a vector of length dim that records the size $m_I$ of each group.

It can be generated by the following methods.

Main.FuzzifiED.Basis — Method

function Basis(cfs :: Confs, qnz_s :: Vector{ComplexF64}, cyc :: Vector{Int64}, perm_o :: Vector{Vector{Int64}}, ph_o :: Vector{Vector{Int64}}, fac_o :: Vector{Vector{ComplexF64}}) :: Basis

generates the basis that respects the discrete symmetries from the configurations that respects the conserved quantities. The discrete $\mathbb{Z}_n$ symmetries are in the form of

\[\mathscr{Z}:\ c_o\to \alpha_o^* c^{(p_o)}_{\pi_o},\quad c_o^\dagger\to \alpha_o c^{(1-p_o)}_{\pi_o}\]

where we use a notation $c^{(1)}=c^\dagger$ and $c^{0}=c$ for convenience, where $\pi_o$ is a permutation of $1,\dots N_o$, $\alpha_o$ is a coefficient, and $p_o$ specified whether or not particle-hole transformation is performed for the orbital. Note that one must guarentee that all these transformations commute with each other and also commute with the conserved quantities.

Arguments

cfs :: Confs is the set of configurations with only conserved quantities ;

qnz_s :: Vector{ComplexF64} is a vector of length the same as the number of discrete symmetries that records the eigenvalue of each transformation ;

cyc :: Vector{Int64} records the cycle of each transformation. For $\mathbb{Z}_n$ symmetry, record $n$ ;

perm_o :: Vector{Vector{Int64}} records the permutation $\pi_o$. It has $N_Z$ elements and each of its elements is a vector of length $N_o$.

ph_o :: Vector{Vector{Int64}} records $p_o$ to determine whether or not to perform a particle-hole transformation. It has $N_Z$ elements and each of its elements is a vector of length $N_o$.

fac_o :: Vector{Vector{ComplexF64}} records the factor $p_o$ is determine whether or not to perform a particle-hole transformation. Each of its elements is a vector of length $N_o$.

Output

bs :: Basis is the resulting Basis object

Main.FuzzifiED.Basis — Method

function Basis(cfs :: Confs) :: Basis

Generate a basis from the configurations without applying the $\mathbb{Z}_2$ symmetries.

Arguments

cfs :: Confs is the set of configurations with only conserved quantities ;

Output

bs :: Basis is the resulting Basis object

The look-back of the weight of a configuration in a state can be done by

Main.FuzzifiED.GetConfWeight — Method

GetConfWeight(bs :: Basis, st :: Vector{ComplexF64}, cf :: Int64) :: ComplexF64
GetConfWeight(bs :: Basis, st :: Vector{Float}, cf :: Int64) :: ComplexF64

looks up a the weight of a configuration in a state.

Arguments

bs :: Basis is the basis of the state ;

st :: Vector{ComplexF64} or st :: Vector{Float64} is a vector of length bs.dim that stores the state ;

cf :: Int64 stores the configuration to be looked-up expressed in a binary number. If the o-1-th bit of conf[i] is 1, then the o-th orbital in the i-th configuration is occupied ; if the bit is 0, then the orbital is empty.

Output

The weight of the configuration in the state

Term

Main.FuzzifiED.Term — Type

mutable struct Term

A Term object records a term that looks like $Uc^{(p_1)}_{o_1}c^{(p_2)}_{o_2}\dots c^{(p_l)}_{o_l}$ in an operator

Fields

coeff :: ComplexF64 records the coefficient $U$

cstr :: Vector{Int64} is a length-$2l$ vector $(p_1,o_1,p_2,o_2,\dots p_l,o_l)$ recording the operator string

Method

It can be generated by the function

function Term(coeff :: ComplexF64, cstr :: Vector{Int64}) :: ComplexF64

The product of terms with a number, the sum and product of terms and the adjoint is defined

Base.:* — Method

function *(fac :: Number, tms :: Vector{Term}) :: Vector{Term}
function -(tms :: Vector{Term}) :: Vector{Term}
function *(tms :: Vector{Term}, fac :: Number) :: Vector{Term}
function /(tms :: Vector{Term}, fac :: Number) :: Vector{Term}

Return the product of a collection of terms with a number.

Base.:+ — Method

function +(tms1 :: Vector{Term}, tms2 :: Vector{Term}) :: Vector{Term}
function -(tms1 :: Vector{Term}, tms2 :: Vector{Term}) :: Vector{Term}

Return the naive sum of two series of terms by taking their union.

Base.:* — Method

function *(tms1 :: Vector{Term}, tms2 :: Vector{Term}) :: Vector{Term}

Return the naive product of two series of terms. The number of terms equals the product of the number of terms in tms1 and tms2. For each term in tms1 $Uc^{(p_1)}_{o_1}\dots$ and tms2 $U'c^{(p'_1)}_{o'_1}\dots$, a new term is formed by taking $UU'c^{(p_1)}_{o_1}\dots c^{(p'_1)}_{o'_1}\dots$

Base.adjoint — Method

function adjoint(tm :: Term) :: Term
function adjoint(tms :: Vector{Term}) :: Vector{Term}

Return the Hermitian conjugate of a series of terms. For each term $Uc^{(p_1)}_{o_1}c^{(p_2)}_{o_2}\dots c^{(p_l)}_{o_l}$, the adjoint is $\bar{U}c^{(1-p_l)}_{o_l}\dots c^{(1-p_2)}_{o_2}c^{(1-p_1)}_{o_1}$

Operator

Main.FuzzifiED.Operator — Type

mutable struct Operator

An Operator object records the sum of terms together with information about its symmetry and the basis of the state it acts on and the basis of the resulting state.

Fields

bsd :: Basis is the basis of the initial state ;

bsf :: Basis is the basis of the final state ;

red_q :: Int64 is a flag that records whether or not the conversion to a sparse martrix can be simplified : if bsd and bsf have exactly the same quantum numbers, and the operator fully respects the symmetries, and all the elements in bsd.cffac and bsf.cffac has the same absolute value, then red_q = 1 ; otherwise red_q = 0 ;

sym_q :: Int64 records the symmetry of the operator : if the matrix is Hermitian, then sym_q = 1 ; if it is symmetric, then sym_q = 2 ; otherwise sym_q = 0 ;

ntm :: Int64 is the number of terms ;

nc :: Int64 is the maximum number of operators in an operator string

cstrs :: Matrix{Int64} is a matrix recording the operator string of each term. Each column corresponds to a term and is padded to the maximum length with -1's.

coeffs :: Vector{ComplexF64} corresponds to the coefficients in each term.

It can be generated by the following methods.

Main.FuzzifiED.Operator — Method

function Operator(bsd :: Basis, bsf :: Basis, terms :: Vector{Term} ; red_q :: Int64, sym_q :: Int64) :: Operator

generates an operator object from a series of terms.

Arguments

bsd :: Basis is the basis of the initial state ;

bsf :: Basis is the basis of the final state ;

terms :: Vector{Term} records the terms ;

red_q :: Int64 is a flag that records whether or not the conversion to a sparse martrix can be simplified : if bsd and bsf have exactly the same quantum numbers, and the operator fully respects the symmetries, and all the elements in bsd.cffac and bsf.cffac has the same absolute value, then red_q = 1 ; otherwise red_q = 0 ;

sym_q :: Int64 records the symmetry of the operator : if the matrix is Hermitian, then sym_q = 1 ; if it is symmetric, then sym_q = 2 ; otherwise sym_q = 0.

The product of an operator on a state and the inner product of a final state, an operator and an initial state can be calculated by

Base.:* — Method

function *(op :: Operator, st_d :: Vector{ComplexF64}) :: Vector{ComplexF64}
function *(op :: Operator, st_d :: Vector{Float64}) :: Vector{Float64}

Measure the action of an operator on a state. st_d must be of length op.bsd.dim. Returns a vector of length op.bsf.dim that represents the final state.

Base.:* — Method

function *(st_fp :: LinearAlgebra.Adjoint{ComplexF64, Vector{ComplexF64}}, op :: Operator, st_d :: Vector{ComplexF64}) :: ComplexF64
function *(st_fp :: LinearAlgebra.Adjoint{Float64, Vector{Float64}}, op :: Operator, st_d :: Vector{Float64}) :: Float64

Measuring the inner product between two states and an operator. st_d must be of length op.bsd.dim and st_fp must be of length op.bsf.dim, and st_fp must be an adjoint.

Sparse matrix

Main.FuzzifiED.OpMat — Type

mutable struct OpMat{Complex64}
mutable struct OpMat{Float64}

This type stores a sparse matrix in the same form as SparseMatrixCSC in SparseArrays. If the matrix is Hermitian or symmetric, only the lower triangle is stored.

Fields

dimd :: Int64 and dimf :: Int64 are the number of columns and rows of the matrix ;

symq :: Int64 records whether or not the matrix is Hermitian or symmetric ;

nel :: Int64 records the number of elements ;

colptr :: Vector{Int64}, rowid :: Vector{Int64} and elval :: Vector{ComplexF64} records the elements of the sparse matrix as in the SparseMatrixCSC elements of Julia.

It can be generated by the following methods.

Main.FuzzifiED.OpMat — Method

function OpMat(op :: Operator ; type :: DataType = ComplexF64) :: OpMat{type}

Generates the sparse matrix from the operator

Arguments

op :: Operator is the operator ;

type :: DataType specifies the type of the matrix. It can either be ComplexF64 or Float64.

The lowest eigenstates of the sparse matrix can be calculated by

Main.FuzzifiED.GetEigensystem — Method

function GetEigensystem(mat :: OpMat{ComplexF64}, nst :: Int64 ; tol :: Float64, ncv :: Int64) :: Tuple{Vector{ComplexF64}, Matrix{ComplexF64}}
function GetEigensystem(mat :: OpMat{Float64}, nst :: Int64 ; tol :: Float64, ncv :: Int64) :: Tuple{Vector{Float64}, Matrix{Float64}}

calls the Arpack package to calculate the lowest eigenstates of sparse matrix.

Arguments

mat :: OpMat{ComplexF64} is the matrix ;

nst :: Int64 is the number of eigenstates to be calculated ;

tol :: Float64 is the tolerence for the Arpack process. The default value is 1E-8 ;

ncv :: Int64 is an euxiliary parameter needed in the Arpack process. The default value is max(2 * nst, nst + 10)

Output

A length-nst array that has the same type as mat recording the eigenvalues, and

A dimd*nst matrix that has the same type as mat where every column records an eigenstate.

The product of an operator on a state and the inner product of a final state, an operator and an initial state can be calculated by

Base.:* — Method

function *(mat :: OpMat{ComplexF64}, st_d :: Vector{ComplexF64}) :: Vector{ComplexF64}
function *(mat :: OpMat{Float64}, st_d :: Vector{Float64}) :: Vector{Float64}

Measure the action of a sparse matrix on a state. st_d must be of length mat.dimd. Returns a vector of length mat.dimf that represents the final state.

Base.:* — Method

function *(st_fp :: LinearAlgebra.Adjoint{ComplexF64, Vector{ComplexF64}}, mat :: OpMat{ComplexF64}, st_d :: Vector{ComplexF64}) :: ComplexF64
function *(st_fp :: LinearAlgebra.Adjoint{Float64, Vector{Float64}}, mat :: OpMat{Float64}, st_d :: Vector{Float64}) :: Float64

Measuring the inner product between two states and a sparse matrix. st_d must be of length mat.dimd and st_fp must be of length mat.dimf, and st_fp must be an adjoint.

Note that sometimes it is needed to transform a state from one basis to another. This can be done by constructing an identity operator.

stf = Operator(bsd, bsf, [Term(1., [-1, -1])]) * std

The OpMat object can be converted with the SparseMatrixCSC object in the SparseArrays package.

Main.FuzzifiED.SparseMatrixCSCFromOpMat — Method

SparseMatrixCSCFromOpMat(mat :: OpMat{ComplexF64}) :: SparseMatrixCSC{Int64,ComplexF64}
SparseMatrixCSCFromOpMat(mat :: OpMat{Float64}) :: SparseMatrixCSC{Int64,Float64}

converts the OpMat objects to a SparseMatrixCSC object in the SparseArrays package.

Main.FuzzifiED.OpMat — Method

OpMat(matcsc :: SparseMatrixCSC{Int64,ComplexF64}) :: OpMat{ComplexF64}
OpMat(matcsc :: SparseMatrixCSC{Int64,Float64}) :: OpMat{Float64}

converts the SparseMatrixCSC object in the SparseArrays package to an OpMat objects.

ITensors support

This package also supports convertion with the Site and OpSum objects from ITensors library and management of DMRG sweeping process. To use these functions, include

using ITensors

at the heading of the Julia script.

Format conversion

The Sites objects in ITensors can be converted to a Confs object with the conserved quantities extracted.

Main.FuzzifiED.ConfsFromSites — Method

function ConfsFromSites(sites :: Vector{Index{Vector{Pair{QN, Int64}}}}, qn_s :: QN) :: Confs

Converts a Sites object in the ITensors package to the Confs object

Arguments

sites :: Vector{Index{Vector{Pair{QN, Int64}}}} is a Sites object. Only Fermion site type is supported, and the quantum numbers of the 0 state must be all zero. Note that this will subject to the limitation in ITensors that the number of conserved quantities must be less than 4.

qn_s :: QN is a QN object that specifies the the quantum number of the selected configuration.

Main.FuzzifiED.ConfsFromSites — Method

function ConfsFromSites(sites :: Vector{Index{Vector{Pair{QN, Int64}}}}, cf_ref :: Vector{Int64}) :: Confs

sites :: Vector{Index{Vector{Pair{QN, Int64}}}} is a Sites object, where the modulus-$\pm 1$ quantum numbers will be taken out, and QNs with other modulus will be discarded automatically. Also note that only Fermion site type is supported, and the quantum numbers of the 0 state must be all zero. Note that this will subject to the limitation in ITensors that the number of conserved quantities must be less than 4.

cf_ref :: Vector{Int64}) is a reference configuration composed of 0 and 1 in ITensors format.

Conversely, the Sites objects in ITensors can also be generated from a set of diagonal quantum numbers

Main.FuzzifiED.SitesFromQN — Method

function SitesFromQN(; qnu_o :: Vector{Vector{Int64}}, qnu_name :: Vector{String}, modul :: Vector{Int64})

returns the ITensors Sites object from the information of quantum numbers

Arguments

qnu_o :: Vector{Vector{Int64}} stores the charge of each orbital under each conserved quantity. See Confs for detail.

qnu_name :: Vector{String} stores the name of each quantum number. Facultive, QN1, QN2, ... by default.

modul :: Vector{Int64} stores the modulus of each quantum number. Store 1 if no modulus. Facultive, all 1 by default.

The OpSum objects in ITensors can be converted with the collection of Term's

Main.FuzzifiED.TermsFromOpSum — Method

function TermsFromOpSum(opsum :: Sum{Scaled{ComplexF64, Prod{Op}}}) :: Vector{Term}

Converts a OpSum object in ITensors to a series of terms. Note that the only operators supported are "C", "Cdag" "N" and "I".

Main.FuzzifiED.OpSumFromTerms — Method

function OpSumFromTerms(tms :: Vector{Term}) :: Sum{Scaled{ComplexF64, Prod{Op}}}

Converts a series of terms to OpSum object in ITensors.

Easy sweep

This tool kit facilitates the management of DMRG process. It automatically records the intermediate results and recover these results if a job is stopped and run again on HPC. It also manages the gradual increase of maximal bond dimensions and the determination of convergence by the criteria of energy. These functions require the package ITensorMPOConstruction, which can be installed by

julia> using Pkg; Pkg.add(url="https://github.com/ITensor/ITensorMPOConstruction.jl.git")

Include also

using ITensorMPOConstruction

at the heading of the Julia script.

Main.FuzzifiED.EasySweep — Method

function EasySweep(id :: String, hmt :: MPO, st00 :: MPS ; path :: String, dim_list :: Vector{Int64}, proj :: Vector{String}, e_tol1 :: Float64, e_tol :: Float64, cutoff :: Vector{Float64}, maxdim0 :: Vector{Float64}, noise0 :: Vector{Float64}, noise :: Vector{Int64}, nsweeps :: Int64, weight :: Float64, observer :: AbstractObserver) :: Tuple{Float64, MPS}

Function

This function automatically performs several rounds of DMRG sweeps with increasing bond dimensions. It first checks the file st_$(id).h5 in a specified repository. If the key st_fin exists, it reads the energy and MPS from the file and return the energy and MPS, otherwise it will perform the DMRG process. For each round, it will try to access the results from the key st_d$(dim_i) in st_$(id).h5, where dim_i is either 0 representing the initial round, or an element of array dim_list. If the key exist, it will read the result ; otherwise it will perform the sweeps using SweepOne. For the initial round, it will take the initial state from st00, the maximal bond dimensions from maxdim0, noise from noise0 and record the results in the key E_d0 and st_d0 in st_$(id).h5. For each of the following round, it will take the result from the previous round as the initial state and perform nsweeps sweeps with the bond dimension dim_list[i]. Each round will be stopped if the energy difference is less than e_tol1. The entire process will be stopped if the energy difference between two rounds is less than e_tol or the bond dimension of the result is less than 0.9 times the maximal bond dimension. The projected states will be accessed from the files specified by proj. It will try to access first the states with the same bond dimension as the projected states. If such states do not exist, it will then access the final state. The resulting energy will be written into the key E_fin in the file st_$(id).h5, and the MPS written into st_fin. The function returns a tuple of energy and the final MPS.

Arguments

id :: String is a string identifying the file to which the results will be accessed and written.

hmt :: MPO is an MPO specifying the Hamiltonian.

st00 :: MPS is an MPS specifying the initial state.

path :: String identifies the path where the results will be accessed and stored. Facultive, ./ by default.

dim_list :: Vector{Int64} :: Int64 is a list that specifies the maximal bond dimensions of each round of sweeps starting from the second round. Facultive, [1000,2000,3000,4000,5000,6000] by default

proj :: Vector{String} specifies the name of the states that will be projected. Facultive, empty by default.

e_tol1 :: Float64 specifies the energy tolerence as a criteria to end the round of sweeps for each round of sweeps. Facultive, 1E-6 by default.

e_tol :: Float64 specifies the energy tolerence as a criteria to end the entire process. Facultive, 1E-7 by default.

cutoff :: Vector{Float64} is the cutoff that will be sent into DMRG. Facultive, [1.E-9] by default.

maxdim0 :: Vector{Int64} specifies the maximal bond dimensions of the first round of sweeps. Facultive, [1E-4,3E-5,1E-5,3E-6,1E-6,3E-7] by default.

noise0 :: Vector{Float64} specifies the noise of each sweep in the initial round and will be sent into DMRG. Facultive, [1E-4,3E-5,1E-5,3E-6,1E-6,3E-7] by default.

noise :: Vector{Float64} specifies the noise of each sweep from the second round and will be sent into DMRG. Facultive, [1E-6,1E-7,0] by default.

nsweeps :: Int64 specifies the number of sweeps in each round from the second rounds. Facultive, 10 by default.

weight :: Float64 specifies the weight of projected states and will be sent into DMRG. Facultive, 100.0 by default.

observer :: AbstractObserver specifies the measurement and cutoff condition for each sweep starting from the second round. Facultive, by default the observer will print the energy and cutoff once the energy difference is less than e_tol at each sweep.

Main.FuzzifiED.SweepOne — Method

function SweepOne(id :: String, hmt :: MPO, st0 :: MPS, dim1 :: Int64 ; path :: String, cutoff :: Vector{Float64}, maxdim :: Vector{Int64}, nsweeps :: Int64, noise :: Vector{Float64}, proj :: Vector{String}, e_tol :: Float64, weight :: Float64, observer :: AbstractObserver) :: Tuple{Float64, MPS}

Function

This function performs one round of nsweeps sweeps. It first checks the file st_$(id).h5 in a specified repository. If the key st_d$(dim1) exists, it reads the energy and MPS from the file and return the energy and MPS, otherwise it will perform the DMRG process with the maximal bond dimension specified by maxdim if it exists, or dim1. The projected states will be read from the key st_d$(dim1) if it exists or st_fin in the file st_$(fi).h5 in the same repository for each string fi in the array proj. The sweeps will be ended if the energy difference is less than etol or whatever criteria is given in observer. The resulting energy will be written into the key E_d$(dim1) in the file st_$(id).h5, and the MPS written into st_d$(dim1). The function returns a tuple of energy and the final MPS.

Arguments

id :: String is a string identifying the file to which the results will be accessed and written.

hmt :: MPO is an MPO specifying the Hamiltonian.

st0 :: MPS is an MPS specifying the initial state.

dim1 :: Int64 is a bond dimension that will be used to identify the result.

path :: String identifies the path where the results will be accessed and stored. Facultive, ./ by default.

cutoff :: Vector{Float64} is the cutoff that will be sent into DMRG. Facultive, [1.E-9] by default.

maxdim :: Vector{Int64} specifies the maximal bond dimension of each sweep. Facultive, [dim1] by default.

nsweeps :: Int64 specifies the number of sweeps in the round. Facultive, 10 by default.

noise :: Vector{Float64} specifies the noise of each sweep and will be sent into DMRG. Facultive, [1E-6,1E-7,0] by default.

proj :: Vector{String} specifies the name of the states that will be projected. Facultive, empty by default.

e_tol :: Float64 specifies the energy tolerence as a criteria to end the sweeps. Facultive, 1E-6 by default.

weight :: Float64 specifies the weight of projected states and will be sent into DMRG. Facultive, 100.0 by default.

observer :: AbstractObserver specifies the measurement and cutoff condition for each sweep. Facultive, by default the observer will print the energy and cutoff once the energy difference is less than e_tol at each sweep.

Main.FuzzifiED.GetMpoSites — Method

function GetMpoSites(id :: String, tms :: Union{Vector{Term}, Sum{Scaled{ComplexF64, Prod{Op}}}} ; path :: String, qnu_o :: Vector{Vector{Int64}}, qnu_name :: Vector{String}, modul :: Vector{Int64}, old = false) :: Tuple{MPO, Vector{<:Index}}

Function

This function returns the MPO and sites for a given operator and a Hilbert space with given quantum numbers. It first checks the file op_$(id).h5 in a specified repository. If the file exists, it will try to read the fields mpo and sites and return the MPO and Sites. Otherwise it will first generates the sites with the quantum numbers given in qnu_o, qnu_name and modul (these objects are often results of a function named Get*Qnu). Then it will generate the MPO with the terms of the operator given in tms. The MPO and sites will be written into the file op_$(id).h5 in the fields mpo and sites.

Arguments

id :: String is a string identifying the file to which the results will be accessed and written.

tms :: Vector{Term} or tms :: Sum{Scaled{ComplexF64, Prod{Op}}} is either an array of terms or a OpSum objects that specifies the expression of the operator.

path :: String identifies the path where the results will be accessed and stored. Facultive, ./ by default.

qnu_o :: Vector{Vector{Int64}} is a list where each element specifies a quantum number. Each element is a list that specifies the charges of each orbital under the quantum number. Obligatory.

qnu_name :: Vector{String} specifies the name of each quantum number. Facultive, QN1, QN2, ... by default.

module :: Vector{Int64} specifies the modulus of each quantum number. Facultive, all 1 by default.

old :: Bool specifies the method of MPO generation. If true, then the MPO() method in ITensors will be used ; if false, then MPO_new() method in ITensorMPOConstruction package will be used. Facultive, false by default.

Main.FuzzifiED.GetMpo — Method

function GetMpo(id :: String, tms :: Union{Vector{Term}, Sum{Scaled{ComplexF64, Prod{Op}}}} ; path :: String, qnu_o :: Vector{Vector{Int64}}, qnu_name :: Vector{String}, modul :: Vector{Int64}, old = false) :: Tuple{MPO, Vector{<:Index}}

Function

This function returns the MPO for a given operator and a given set of sites. It first checks the file op_$(id).h5 in a specified repository. If the file exists, it will try to read the fields mpo and return the MPO it has read. Otherwise it will generate the MPO with the terms of the operator given in tms. The MPO will be written into the file op_$(id).h5 in the fields mpo.

Arguments

id :: String is a string identifying the file to which the results will be accessed and written.

tms :: Vector{Term} or tms :: Sum{Scaled{ComplexF64, Prod{Op}}} is either an array of terms or a OpSum objects that specifies the expression of the operator.

sites :: Vector{<:Index} specifies the sites that the operator is acting on.

path :: String identifies the path where the results will be accessed and stored. Facultive, ./ by default.

old :: Bool specifies the method of MPO generation. If true, then the MPO() method in ITensors will be used ; if false, then MPO_new() method in ITensorMPOConstruction package will be used. Facultive, false by default.

Built-in models

Main.FuzzifiED.GetIntMatrix — Method

function GetIntMatrix(nm :: Int64, ps_pot :: Vector{Number}) :: Array{ComplexF64, 3}

Argument

nm :: Int64 is the number of orbitals

ps_pot :: Vector{Number} is the vector of non-zero pseudopotentials

Output

A nm*nm*nm array giving the interaction matrix $U_{m_1,m_2,m_3,-m_1-m_2-m_3}$

Main.FuzzifiED.GetL2Terms — Method

function GetL2Terms(nm :: Int64, nf :: Int64) :: Vector{Term}

Return the terms for the total angular momentum.

Arguments

nm :: Int64 is the number of orbitals.

nf :: Int64 is the number of flavours.

Ising model

Main.FuzzifiED.GetIsingQnu — Method

function GetIsingQnu(nm :: Int64) :: @NamedTuple{qnu_o :: Vector{Vector{Int64}}, qnu_name :: Vector{String}, modul :: Vector{Int64}}

returns the diagonal quantum numbers, i.e., particle number $N_e$ and angular momentum $L_z$, of the fuzzy sphere Ising model.

Arguments

nm :: Int64 is the number of orbitals ;

Output

A named tuple with three elements that can be directly fed into SitesFromQN

qnu_o :: Vector{Vector{Int64}} stores the charge of each orbital under each conserved quantity. See Confs for detail.

qnu_name :: Vector{String} stores the name of each quantum number.

modul :: Vector{Int64} stores the modulus of each quantum number, 1 if no modulus.

Main.FuzzifiED.GetIsingConfs — Method

function GetIsingConfs(nm :: Int64 ; ne :: Int64 = 0, lz :: Float64) :: Confs

Return the configurations with conserved particle number $N_e$ and angular momentum $L_z$.

Arguments

nm :: Int64 is the number of orbitals $2s+1$.

ne :: Int64 is the number of electrons.

lz :: Float64 is the angular momentum. Facultive, 0 by default.

Main.FuzzifiED.GetIsingBasis — Method

function GetIsingBasis(cfs :: Confs ; qn_p :: Int64 = 0, qn_z :: Int64 = 0, qn_r :: Int64 = 0) :: Basis

Return the basis with conserved parity $\mathscr{P}$, flavour symmetry $\mathscr{Z}$ and $\pi$-rotation along $y$-axis $\mathscr{R}$ from the configurations already generated. Quantum numbers set to zero signify that they are not conserved.

Arguments

cfs :: Confs is the configurations generated by GetIsingConfs

qn_p :: Int64 is quantum number for parity transformation. Facultive, 0 by default.

qn_z :: Int64 is the particle quantum number for $\mathbb{Z}_2$-flavour transformation. Facultive, 0 by default.

qn_r :: Int64 is the quantum number for $\pi$ rotation along $y$-axis compared with the ground state. Facultive, 0 by default.

Main.FuzzifiED.GetIsingIntTerms — Method

function GetIsingIntTerms(nm :: Int64, ps_pot :: Vector) :: Vector{Term}

Returns the terms for the ising interaction

\[\sum_{m_1m_2m_3m_4}2U_{m_1m_2m_3m_4}c^{\dagger}_{m_1\uparrow}c^{\dagger}_{m_2\downarrow}c_{m_3\downarrow}c_{m_4\uparrow}\]

from the pseudopotentials.

Arguments

nm :: Int64 is the number of orbitals $2s+1$.

ps_pot :: Vector{Number} is the pseudopotential of Ising interaction.

Main.FuzzifiED.GetXPolTerms — Method

function GetXPolTerms(nm :: Int64)

Returns the terms for the density operator $n^x_{l=0,m=0}$

Arguments

nm :: Int64 is the number of orbitals.

Main.FuzzifiED.GetZPolTerms — Method

function GetZPolTerms(nm :: Int64)

Returns the terms for the density operator $n^z_{l=0,m=0}$

Arguments

nm :: Int64 is the number of orbitals.

Ising model in the X basis

Main.FuzzifiED.GetIsingXQnu — Method

function GetIsingQnu(nm :: Int64) :: @NamedTuple{qnu_o :: Vector{Vector{Int64}}, qnu_name :: Vector{String}, modul :: Vector{Int64}}

returns the diagonal quantum numbers, i.e., particle number $N_e$, angular momentum $L_z$ and spin-parity $S_z$, of the fuzzy sphere Ising model. The Ising model is written in the basis of

\[c_+^\dagger=\frac{c_\uparrow^\dagger+c_\downarrow^\dagger}{\sqrt{2}},\quad c_-^\dagger=\frac{c_\uparrow^\dagger-c_\downarrow^\dagger}{\sqrt{2}}\]

and the parity of the $\mathbb{Z}_2$-odd particle numbers is conserved.

Arguments

nm :: Int64 is the number of orbitals ;

Output

A named tuple with three elements that can be directly fed into SitesFromQN

qnu_o :: Vector{Vector{Int64}} stores the charge of each orbital under each conserved quantity. See Confs for detail.

qnu_name :: Vector{String} stores the name of each quantum number.

modul :: Vector{Int64} stores the modulus of each quantum number, 1 if no modulus.

Main.FuzzifiED.GetIsingXConfs — Method

function GetIsingConfs(nm :: Int64 ; ne :: Int64 = 0, lz :: Float64) :: Confs

Return the configurations with conserved particle number $N_e$, angular momentum $L_z$ and spin-parity $S_z$ of the fuzzy sphere Ising model.

Arguments

nm :: Int64 is the number of orbitals $2s+1$.

ne :: Int64 is the number of electrons.

lz :: Float64 is the angular momentum. Facultive, 0 by default.

Main.FuzzifiED.GetIsingXIntTerms — Method

function GetIsingIntTerms(nm :: Int64, ps_pot :: Vector) :: Vector{Term}

Returns the terms for the ising interaction.

\[\sum_{m_1m_2m_3m_4}2U_{m_1m_2m_3m_4}c^{\dagger}_{m_1\uparrow}c^{\dagger}_{m_2\downarrow}c_{m_3\downarrow}c_{m_4\uparrow}\]

written in the basis of $c_\pm$ from the pseudopotentials.

Arguments

nm :: Int64 is the number of orbitals $2s+1$.

ps_pot :: Vector{Number} is the pseudopotential of Ising interaction.

$\mathrm{Sp}(N)$ model

Main.FuzzifiED.GetSpnConfs — Method

function GetSpnConfs(nm :: Int64, nf :: Int64, ne :: Int64 ; lz :: Float64 = 0.0, sz :: Vector{Int64}) :: Confs

Return the configurations with conserved particle number $N_e$, angular momentum $L_z$ and Cartans of the $\mathrm{Sp}(N)$ symmetry

\[\begin{aligned} N_e&=\sum_{mf}n_{mf}\\ L_z&=\sum_{mf}mn_{mf}\\ S_{z,i}&=\sum_m (n_{m,i}-n_{m,i+N_f/2})&i&=1,\dots,N_f/2 \end{aligned}\]

Arguments

nm :: Int64 is the number of orbitals.

nf :: Int64 is the number of flavours.

ne :: Int64 is the number of electrons.

lz :: Float64 is the angular momentum. Facultive, 0 by default.

sz :: Vector{Int64} is $N_f/2$ numbers that give $\mathrm{Sp}(N)$ Cartan. Facultive, a vector of $N_f/2$ zeros by default.

Main.FuzzifiED.GetSpnBasis — Method

function GetSpnBasis(cfs :: Confs, nf :: Int64 ; qn_p :: Int64 = 0, qn_r :: Int64 = 0, qn_z :: Vector{Int64}, qn_x :: Vector{Int64} :: Basis

Return the basis with conserved parity $\mathscr{P}$, $\pi$-rotation along $y$-axis $\mathscr{R}$, and flavour symmetries $\mathscr{Z}_i,\mathscr{X}_i$ from the configurations already generated. Quantum numbers set to zero signify that they are not conserved.

\[\begin{aligned} \mathscr{P}:c^{\dagger}_{mf}&\mapsto c_{m,f+N_f/2},&c^{\dagger}_{m,f+N_f/2}&\mapsto-c_{m,f}\\ \mathscr{R}:c^{\dagger}_{mf}&\mapsto c^{\dagger}_{-m,f}\\ \mathscr{Z}_i:c^{\dagger}_{mi}&\mapsto c^{\dagger}_{m,i+N_f/2}&c^{\dagger}_{m,i+N_f/2}&\mapsto-c^{\dagger}_{m,i},&i=1,2,\dots,N_f/2\\ \mathscr{X}_i:c^{\dagger}_{m,2i-1}&\leftrightarrow c^{\dagger}_{m,2i}&c^{\dagger}_{m,2i-1+N_f/2}&\leftrightarrow c^{\dagger}_{m,2i+N_f/2},&i=1,2,\dots,N_f/4\\ \end{aligned}\]

Note that sometimes some $\mathscr{Z}_i$ and $\mathscr{X}_i$ cannot be implemented at the same time.

Arguments

cfs :: Confs is the configurations generated by GetIsingConfs.

nf :: Int64 is the number of flavours.

qn_p :: Int64 is quantum number for parity transformation. Facultive, 0 by default.

qn_r :: Int64 is the quantum number for $\pi$ rotation along $y$-axis compared with the ground state. Facultive, 0 by default.

qn_z :: Vector{Int64} is $N_f/2$ numbers that give $\mathscr{Z}_i$ flavour symmetry. Facultive, all 0 by default.

qn_x :: Vector{Int64} is $N_f/4$ numbers that give $\mathscr{X}_i$ flavour symmetry. Facultive, all 0 by default.

Main.FuzzifiED.GetIdDenIntTerms — Method

function GetIdDenIntTerms(nm :: Int64, nf :: Int64, ps_pot :: Vector) :: Vector{Term}

Returns the terms for the Hubbard density-density interaction

\[\sum_{m_1m_2m_3m_4ff'}2U_{m_1m_2m_3m_4}c^{\dagger}_{m_1f}c^{\dagger}_{m_2f'}c_{m_3f'}c_{m_4f}\]

from the pseudopotentials.

Arguments

nm :: Int64 is the number of orbitals $2s+1$.

nf :: Int64 is the number of flavours.

ps_pot :: Vector{Number} is the pseudopotential of the hubbard interaction.

Main.FuzzifiED.GetSpnPairIntTerms — Method

function GetSpnPairIntTerms(nm :: Int64, nf :: Int64, ps_pot :: Vector) :: Vector{Term}

Returns the terms for the $\mathrm{Sp}(N)$ pair-pair interaction

\[\sum_{m_1m_2m_3m_4ff'}2U_{m_1m_2m_3m_4}c^{\dagger}_{m_1f}c^{\dagger}_{m_2,f+N_f/2}c_{m_3 f'+N_f/2}c_{m_4f'}\]

from the pseudopotentials.

Arguments

nm :: Int64 is the number of orbitals.

nf :: Int64 is the number of flavours.

ps_pot :: Vector{Number} is the pseudopotential of the hubbard interaction.

Main.FuzzifiED.GetSpnC2Terms — Method

function GetSpnC2Terms(nm :: Int64, nf :: Int64) :: Vector{Term}

Returns the quadratic Casimir $C_2$ of the $\mathrm{Sp}(N)$ symmetry.

Arguments

nm :: Int64 is the number of orbitals.

nf :: Int64 is the number of flavours.