Built-in models

General models on fuzzy sphere

Main.FuzzifiED.GetIntMatrix — Method

function GetIntMatrix(nm :: Int64, ps_pot :: Vector{Number}) :: Array{ComplexF64, 3}

Argument

nm :: Int64 is the number of orbitals

ps_pot :: Vector{Number} is the vector of non-zero pseudopotentials

Output

A nm*nm*nm array giving the interaction matrix $U_{m_1,m_2,m_3,-m_1-m_2-m_3}$

Main.FuzzifiED.GetL2Terms — Method

function GetL2Terms(nm :: Int64, nf :: Int64) :: Vector{Term}

Return the terms for the total angular momentum.

Arguments

nm :: Int64 is the number of orbitals.

nf :: Int64 is the number of flavours.

Main.FuzzifiED.GetDenIntTerms — Method

GetDenIntTerms(nm :: Int64, nf :: Int64 ; ps_pot :: Vector{<:Number}, mat_a :: Matrix{<:Number}, mat_b :: Matrix{<:Number}) :: Vector{Term}

Return the normal-ordered density-density term in the Hamiltonian

\[∑_{\{m_i,f_i\}}U_{m_1m_2m_3m_4}M^A_{f_1f_4}M^B_{f_2f_3}c^{†}_{m_1f_1}c^{†}_{m_2f_2}c_{m_3f_3}c_{m_4f_4}\]

Arguments

nm :: Int64 is the number of orbitals.

nf :: Int64 is the number of flavours.

ps_pot :: Vector{<:Number} is a list of numbers specifying the pseudopotentials for the interacting matrix $U_{m_1m_2m_3m_4}$. Facultative, [1.0] by default.

mat_a :: Matrix{<:Number} is a nf*nf matrix specifying $M^A_{ff'}$. Facultative, $I_{N_f}$ by default.

mat_b :: Matrix{<:Number} is a nf*nf matrix specifying $M^B_{ff'}$. Facultative, the Hermitian conjugate of mat_a by default.

Main.FuzzifiED.GetPairIntTerms — Method

GetPairIntTerms(nm :: Int64, nf :: Int64 ; ps_pot :: Vector{<:Number}, mat_a :: Matrix{<:Number}, mat_b :: Matrix{<:Number}) :: Vector{Term}

Return the normal-ordered pair-pair interaction term in the Hamiltonian

\[∑_{\{m_i,f_i\}}U_{m_1m_2m_3m_4}M^A_{f_1f_2}M^B_{f_4f_3}c^{†}_{m_1f_1}c^{†}_{m_2f_2}c_{m_3f_3}c_{m_4f_4}\]

Arguments

nm :: Int64 is the number of orbitals.

nf :: Int64 is the number of flavours.

ps_pot :: Vector{<:Number} is a list of numbers specifying the pseudopotentials for the interacting matrix $U_{m_1m_2m_3m_4}$. Facultative, [1.0] by default.

mat_a :: Matrix{<:Number} is a nf*nf matrix specifying $M^A_{ff'}$. Facultative, $I_{N_f}$ by default.

mat_b :: Matrix{<:Number} is a nf*nf matrix specifying $M^B_{ff'}$. Facultative, the Hermitian conjugate of mat_a by default.

Main.FuzzifiED.GetPolTerms — Method

GetPolTerms(nm :: Int64, nf :: Int64 ; mat :: Matrix{<:Number} = Matrix{Float64}(I, nf, nf)) :: Vector{Term}

Return the polarisation term in the Hamiltonian

\[∑_{mff'}c^{†}_{mf}M_{ff'}c_{mf'}\]

Arguments

nm :: Int64 is the number of orbitals ;

nf :: Int64 is the number of flavours ;

mat :: Matrix{<:Number} is a nf*nf matrix specifying $M_{ff'}$. Facultative, $I_{N_f}$ by default.

Spherical observables

FuzzifiED supports local observables on sphere that can be decomposed into angular components $\Phi(\Omega)=\sum_{lm}\Phi_{lm}Y^{(s)}_{lm}$

Main.FuzzifiED.SphereObs — Type

mutable struct SphereObs

stores the information of a local observable (or local operator) $\Phi$ that can be decomposed into angular components.

\[ \Phi(\Omega)=∑_{lm}\Phi_{lm}Y^{(s)}_{lm}\]

Fields

s2 :: Int64 is twice the spin $2s$ of the observable.

l2m :: Int64 is twice the maximal angular momentum $2l_{\max}$ of the components of the observable.

get_comp :: Function is a function get_comp(l2 :: Int64, m2 :: Int64) :: Vector{Term} that sends the component specified by a tuple of integers $(2l,2m)$ where $|s|\leq l\leq l_{\max}, -l\leq m\leq l$ to a list of terms that specifies the expression of the component.

stored_q :: Bool is a boolean that specifies whether or not each component of the observable is stored.

comps :: Dict{Tuple{Int64, Int64}, Vector{Term}} stores each component of the observable in the format of a dictionary whose keys are the tuples of integers $(2l,2m)$ and values are the lists of terms that specifies the expression of the component.

It can be initialised with the following methods

Main.FuzzifiED.SphereObs — Method

function SphereObs(s2 :: Int64, l2m :: Int64, get_comp :: Function) :: SphereObs

initialises the observable from $2s$, $2l_{\max}$ and the function $(l,m)↦\Phi_{lm}$

Arguments

s2 :: Int64 is twice the spin $2s$ of the observable.

l2m :: Int64 is twice the maximal angular momentum $2l_{\max}$ of the components of the observable.

get_comp :: Function is a function get_comp(l2 :: Int64, m2 :: Int64) :: Vector{Term} that sends the component specified by a tuple of integers $(2l,2m)$ where $|s|\leq s\leq l_{\max}, -l\leq m\leq l$ to a list of terms that specifies the expression of the component.

Main.FuzzifiED.SphereObs — Method

function SphereObs(s2 :: Int64, l2m :: Int64, get_comp :: Function) :: SphereObs

initialises the observable from $2s$, $2l_{\max}$ and a list of $\Phi_{lm}$ specified by a dictionary.

Arguments

s2 :: Int64 is twice the spin $2s$ of the observable.

l2m :: Int64 is twice the maximal angular momentum $2l_{\max}$ of the components of the observable.

comps :: Dict{Tuple{Int64, Int64}, Vector{Term}} stores each component of the observable in the format of a dictionary whose keys are the tuples of integers $(2l,2m)$ and values are the lists of terms that specifies the expression of the component.

The following methods explicitly calculates and stores each component

Main.FuzzifiED.StoreComps! — Method

function StoreComps!(obs :: SphereObs) :: Nothing

calculates and stores each component of the observable obs and replace the function in obs by the list of calculated components.

Main.FuzzifiED.StoreComps — Method

function StoreComps(obs :: SphereObs) :: SphereObs

calculates and stores each component of the observable obs and return a new observable with the list of calculated components.

The multiplication, addition, conjugate and Laplacian operation of an observable is supported

Base.:* — Method

function *(fac :: Number, obs :: SphereObs) :: SphereObs
function *(obs :: SphereObs, fac :: Number) :: SphereObs
function /(obs :: SphereObs, fac :: Number) :: SphereObs
function -(obs :: SphereObs) :: SphereObs

enables the multiplication of an observable with a number.

Base.:+ — Method

function +(obs1 :: SphereObs, obs2 :: SphereObs) :: SphereObs
function -(obs1 :: SphereObs, obs2 :: SphereObs) :: SphereObs

enables the addition of two observables.

Base.adjoint — Method

function adjoint(obs :: SphereObs) :: SphereObs

enables the Hermitian conjugate of a spherical observable.

\[\begin{aligned} Φ^†(Ω)&=∑_{lm}(Φ_{lm})^†\bar{Y}^{(s)}_{lm}(Ω)=∑_{lm}(Φ_{lm})^†(-1)^{s+m}Y^{(-s)}_{l,-m}(Ω)\\ (Φ^†)_{lm}&=(-1)^{s-m}(Φ_{l,-m})^† \end{aligned}\]

Base.:* — Method

function *(obs1 :: SphereObs, obs2 :: SphereObs) :: SphereObs

enables the multiplication of two observable by making use of the composition of two monopole harmonics into one.

Main.FuzzifiED.Laplacian — Method

function Laplacian(obs :: SphereObs) :: SphereObs

Takes the Laplacian of an observable

\[ (∇^2Φ)_{lm}=-l(l+1)Φ_{lm}\]

The observables can be evaluated either at an angular component or at a real-space point.

Main.FuzzifiED.GetComponent — Method

function GetComponent(obs :: SphereObs, l :: Number, m :: Number) :: Vector{Term}

returns an angular component $Φ_{lm}$ of an observable in the format of a list of terms.

Main.FuzzifiED.GetPointValue — Method

function GetPointValue(obs :: SphereObs, θ :: Float64, ϕ :: Float64) :: Vector{Term}

evaluates an observable at one point $Φ(θ,ϕ)$ in the format of a list of terms.

Two types of operators, viz. electrons and density operators are built-in.

Main.FuzzifiED.Electron — Method

function Electron(nm :: Int64, nf :: Int64, f :: Int64) :: SphereObs

returns the electron annihilation operator $ψ_f$

Arguments

$nf :: Int64$ is the number of flavours.

$nm :: Int64$ is the number of orbitals.

$f :: Int64$ is the index of the orbital to be taken.

Main.FuzzifiED.Density — Method

function Density(nm :: Int64, nf :: Int64 ; mat :: Matrix{<:Number}) :: SphereObs

returns the density operator $n=∑_{ff'}ψ^†_{f}M_{ff'}ψ_{f'}$

Arguments

$nf :: Int64$ is the number of flavours.

$nm :: Int64$ is the number of orbitals.

$mat :: Int64$ is the matrix $M_{ff'}$. Facultative, identity matrix $\mathbb{I}$ by default.

Ising model and Potts model

The following methods is especially helpful for Ising model and Potts model.

Main.FuzzifiED.GetLzQnu — Method

function GetLzQnu(nm :: Int64, nf :: Int64) :: @NamedTuple{qnu_o :: Vector{Vector{Int64}}, qnu_name :: Vector{String}, modul :: Vector{Int64}}

returns the diagonal quantum numbers, i.e., particle number $N_e$ and angular momentum $L_z+sN_e$

\[\begin{aligned} N_e&=∑_o n_o\\ L_z+sN_e&=∑_{mf}(m+s)n_o \end{aligned}\]

Arguments

nm :: Int64 is the number of orbitals ;

nf :: Int64 is the number of flavours ;

Output

A named tuple with three elements that can be directly fed into SitesFromQnu

qnu_o :: Vector{Vector{Int64}} stores the charge of each orbital under each conserved quantity. See Confs for detail.

qnu_name :: Vector{String} stores the name of each quantum number.

modul :: Vector{Int64} stores the modulus of each quantum number, 1 if no modulus.

Main.FuzzifiED.GetLzConfs — Method

function GetLzConfs(nm :: Int64, nf :: Int64, ne :: Int64 ; lz :: Float64) :: Confs

Return the configurations with conserved particle number $N_e$ and angular momentum $L_z$.

Arguments

nm :: Int64 is the number of orbitals $2s+1$.

nf :: Int64 is the number of flavours ;

ne :: Int64 is the number of electrons.

lz :: Float64 is the angular momentum. Facultative, 0 by default.

Main.FuzzifiED.GetLzZnQnu — Method

function GetLzZnQnu(nm :: Int64, nf :: Int64) :: @NamedTuple{qnu_o :: Vector{Vector{Int64}}, qnu_name :: Vector{String}, modul :: Vector{Int64}}

returns the diagonal quantum numbers, i.e., particle number $N_e$, angular momentum $L_z+sN_e$ and flavour charge $Z_{N_f}$

\[\begin{aligned} N_e&=∑_o n_o\\ L_z+sN_e&=∑_{mf}(m+s)n_{mf}\\ Z_{N_f}&=∑_{m,f=0}^{N_f-1}fn_{mf}\ \mathrm{mod}\ N_f \end{aligned}\]

Arguments

nm :: Int64 is the number of orbitals ;

nf :: Int64 is the number of flavours ;

Output

A named tuple with three elements that can be directly fed into SitesFromQnu

qnu_o :: Vector{Vector{Int64}} stores the charge of each orbital under each conserved quantity. See Confs for detail.

qnu_name :: Vector{String} stores the name of each quantum number.

modul :: Vector{Int64} stores the modulus of each quantum number, 1 if no modulus.

Main.FuzzifiED.GetLzZnConfs — Method

function GetLzZnConfs(nm :: Int64, nf :: Int64, ne :: Int64 ; lz :: Float64, zn :: Int64 = 0) :: Confs

Return the configurations with conserved particle number $N_e$, angular momentum $L_z$ and flavour charge $Z_{N_f}$.

Arguments

nm :: Int64 is the number of orbitals $2s+1$.

nf :: Int64 is the number of flavours ;

ne :: Int64 is the number of electrons.

lz :: Float64 is the angular momentum. Facultative, 0 by default.

zn :: Float64 is the flavour charge. Facultative, 0 by default.

Main.FuzzifiED.GetIsingBasis — Method

function GetIsingBasis(cfs :: Confs ; qn_p :: Int64 = 0, qn_z :: Int64 = 0, qn_r :: Int64 = 0) :: Basis

Return the basis with conserved parity $\mathscr{P}$, flavour symmetry $𝒵$ and $π$-rotation along $y$-axis $\mathscr{R}$ from the configurations already generated. Quantum numbers set to zero signify that they are not conserved.

Arguments

cfs :: Confs is the configurations generated by GetLzConfs or GetLzZnConfs.

qn_p :: Int64 is quantum number for parity transformation. Facultative, 0 by default.

qn_z :: Int64 is the particle quantum number for $ℤ_2$-flavour transformation. Facultative, 0 by default.

qn_r :: Int64 is the quantum number for $π$ rotation along $y$-axis compared with the ground state. Facultative, 0 by default.

Main.FuzzifiED.GetSnBasis — Method

function GetSnBasis(cfs :: Confs, nf :: Int64 ; qn_r :: Int64 = 0, perm :: Vector, qn_z :: Vector{<:Number}) :: Basis

Return the basis where the $π$-rotation along $y$-axis $\mathscr{R}$ and certain permutationss of flavour are implemented. Quantum numbers set to zero signify that they are not conserved.

Arguments

cfs :: Confs is the configurations generated by GetLzConfs or GetLzZnConfs.

nf :: Int64 is the number of flavours

qn_r :: Int64 is the quantum number for $π$ rotation along $y$-axis compared with the ground state. Facultative, 0 by default.

perm :: Vector{Vector{Int64}} is a list where each element specifies a permutation of flavour indices (from 1 to $N_f$) in the format of a cycle. Facultative, empty by default.

qn_z :: Vector{<:Number} is a list where each element specifies the quantum number under the flavour permutation. Facultative, empty by default.

Main.FuzzifiED.GetIsingIntTerms — Method

function GetIsingIntTerms(nm :: Int64 ; ps_pot :: Vector) :: Vector{Term}

Returns the terms for the ising interaction

\[∑_{m_1m_2m_3m_4}2U_{m_1m_2m_3m_4}c^{†}_{m_1\uparrow}c^{†}_{m_2\downarrow}c_{m_3\downarrow}c_{m_4\uparrow}\]

from the pseudopotentials.

Arguments

nm :: Int64 is the number of orbitals $2s+1$.

ps_pot :: Vector{Number} is the pseudopotential of Ising interaction.

Main.FuzzifiED.GetXPolTerms — Method

function GetXPolTerms(nm :: Int64)

Returns the terms for the density operator $n^x_{l=0,m=0}$

Arguments

nm :: Int64 is the number of orbitals.

Main.FuzzifiED.GetZPolTerms — Method

function GetZPolTerms(nm :: Int64)

Returns the terms for the density operator $n^z_{l=0,m=0}$

Arguments

nm :: Int64 is the number of orbitals.

$\mathrm{Sp}(N)$ model

The following methods is especially helpful for $\mathrm{Sp}(N)$ models.

Main.FuzzifiED.GetSpnConfs — Method

function GetSpnConfs(nm :: Int64, nf :: Int64, ne :: Int64 ; lz :: Float64 = 0.0, sz :: Vector{Int64}) :: Confs

Return the configurations with conserved particle number $N_e$, angular momentum $L_z$ and Cartans of the $\mathrm{Sp}(N)$ symmetry

\[\begin{aligned} N_e&=∑_{mf}n_{mf}\\ L_z&=∑_{mf}mn_{mf}\\ S_{z,i}&=∑_m (n_{m,i}-n_{m,i+N_f/2})&i&=1,…,N_f/2 \end{aligned}\]

Arguments

nm :: Int64 is the number of orbitals.

nf :: Int64 is the number of flavours.

ne :: Int64 is the number of electrons.

lz :: Float64 is the angular momentum. Facultative, 0 by default.

sz :: Vector{Int64} is $N_f/2$ numbers that give $\mathrm{Sp}(N)$ Cartan. Facultative, a vector of $N_f/2$ zeros by default.

Main.FuzzifiED.GetSpnBasis — Method

function GetSpnBasis(cfs :: Confs, nf :: Int64 ; qn_p :: Int64 = 0, qn_r :: Int64 = 0, qn_z :: Vector{Int64}, qn_x :: Vector{Int64} :: Basis

Return the basis with conserved parity $\mathscr{P}$, $π$-rotation along $y$-axis $\mathscr{R}$, and flavour symmetries $𝒵_i,\mathscr{X}_i$ from the configurations already generated. Quantum numbers set to zero signify that they are not conserved.

\[\begin{aligned} \mathscr{P}&:c^{†}_{mf}↦ c_{m,f+N_f/2},&c^{†}_{m,f+N_f/2}&↦-c_{m,f}\\ \mathscr{R}&:c^{†}_{mf}↦ c^{†}_{-m,f}\\ 𝒵_i&:c^{†}_{mi}↦ c^{†}_{m,i+N_f/2}&c^{†}_{m,i+N_f/2}&↦-c^{†}_{m,i},&i=1,2,…,N_f/2\\ \mathscr{X}_i&:c^{†}_{m,2i-1}\leftrightarrow c^{†}_{m,2i}&c^{†}_{m,2i-1+N_f/2}&\leftrightarrow c^{†}_{m,2i+N_f/2},&i=1,2,…,N_f/4\\ \end{aligned}\]

Note that sometimes some $𝒵_i$ and $\mathscr{X}_i$ cannot be implemented at the same time.

Arguments

cfs :: Confs is the configurations generated by GetSpnConfs.

nf :: Int64 is the number of flavours.

qn_p :: Int64 is quantum number for parity transformation. Facultative, 0 by default.

qn_r :: Int64 is the quantum number for $π$ rotation along $y$-axis compared with the ground state. Facultative, 0 by default.

qn_z :: Vector{Int64} is $N_f/2$ numbers that give $𝒵_i$ flavour symmetry. Facultative, all 0 by default.

qn_x :: Vector{Int64} is $N_f/4$ numbers that give $\mathscr{X}_i$ flavour symmetry. Facultative, all 0 by default.

Main.FuzzifiED.GetSpnPairIntTerms — Method

function GetSpnPairIntTerms(nm :: Int64, nf :: Int64 ; ps_pot :: Vector) :: Vector{Term}

Returns the terms for the $\mathrm{Sp}(N)$ pair-pair interaction

\[∑_{m_1m_2m_3m_4ff'}2U_{m_1m_2m_3m_4}c^{†}_{m_1f}c^{†}_{m_2,f+N_f/2}c_{m_3 f'+N_f/2}c_{m_4f'}\]

from the pseudopotentials.

Arguments

nm :: Int64 is the number of orbitals.

nf :: Int64 is the number of flavours.

ps_pot :: Vector{Number} is the pseudopotential of the hubbard interaction.

Main.FuzzifiED.GetSpnC2Terms — Method

function GetSpnC2Terms(nm :: Int64, nf :: Int64) :: Vector{Term}

Returns the quadratic Casimir $C_2$ of the $\mathrm{Sp}(N)$ symmetry.

Arguments

nm :: Int64 is the number of orbitals.

nf :: Int64 is the number of flavours.